feat: implement SelectCreatable component with feature flag for event type selection
