+++ src/constants/featureFlags.constants.ts
@@ -1,4 +1,5 @@
 export const featureFlags = {
 	displayDiscordIntegration: import.meta.env.DISPLAY_DISCORD_INTEGRATION,
 	displaySlackSocketIntegration: import.meta.env.DISPLAY_SLACK_SOCKET_INTEGRATION,
+	displayComBoxTrigger: import.meta.env.DISPLAY_COMBOX_TRIGGER,
 };
